requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.0
black==23.7.0
flake8==6.1.0

